refactored for do while implementation
A description for the choice of loop is provided at the bottom of the loop underneath the while. As well as this I made use of TryParse instead of Try Catch as it was more readable. ternary was also used to make the invalid input prompt use less lines however, it is arguably less readable.